#88 extract checks for shape transformer for more modularity
