+++ api/parser/src/main/scala/de/htwg/zeta/parser/check/ErrorCheck.scala
@@ -1,0 +1,7 @@
+package de.htwg.zeta.parser.check
+
+import de.htwg.zeta.parser.check.Check.Id
+
+trait ErrorCheck {
+  def check(): List[Id]
+}
